+++ Tests/RunnerLibTests/SPMDangerTests.swift
@@ -6,6 +6,10 @@ final class SPMDangerTests: XCTestCase {
         "testPackage.swift"
     }
 
+    var modernBuildPath: String {
+        "testPath/.build/out/Products/Debug"
+    }
+
     func testItReturnsTrueWhenThePackageHasTheDangerLib() throws {
         let spmDanger = SPMDanger(packagePath: testPackage, readFile: { _ in ".library(name: \"DangerDeps\"" })
 
@@ -100,22 +104,58 @@ final class SPMDangerTests: XCTestCase {
     }
 
     func testItReturnsTheCorrectBuildFolder() throws {
+        let fileManager = StubbedFileManager()
+        fileManager.existingPaths = []
+
         XCTAssertEqual(
             SPMDanger(
                 packagePath: testPackage,
                 readFile: { _ in ".library(name: \"DangerDepsEigen\"" },
-                fileManager: StubbedFileManager()
+                fileManager: fileManager
             )?.buildFolder,
             "testPath/.build/debug"
+        )
+    }
+
+    func testItReturnsTheModernBuildFolderWhenPresent() throws {
+        let fileManager = StubbedFileManager()
+        fileManager.existingPaths = Set([modernBuildPath])
+
+        XCTAssertEqual(
+            SPMDanger(
+                packagePath: testPackage,
+                readFile: { _ in ".library(name: \"DangerDepsEigen\"" },
+                fileManager: fileManager
+            )?.buildFolder,
+            modernBuildPath
+        )
+    }
+
+    func testItReturnsTheModernModuleFolderWhenPresent() throws {
+        let fileManager = StubbedFileManager()
+        fileManager.existingPaths = Set([modernBuildPath])
+
+        XCTAssertEqual(
+            SPMDanger(
+                packagePath: testPackage,
+                readFile: { _ in ".library(name: \"DangerDepsEigen\"" },
+                fileManager: fileManager
+            )?.moduleFolder,
+            modernBuildPath
         )
     }
 }
 
 private class StubbedFileManager: FileManager {
     fileprivate var stubbedFileExists: Bool = true
+    fileprivate var existingPaths: Set<String>?
 
-    override func fileExists(atPath _: String) -> Bool {
-        stubbedFileExists
+    override func fileExists(atPath path: String) -> Bool {
+        if let existingPaths {
+            return existingPaths.contains(path)
+        }
+
+        return stubbedFileExists
     }
 
     override var currentDirectoryPath: String {